requirements.txt
streamlit
google-generativeai
sentence-transformers
chromadb
pandas
python-dotenv